Refactored into separate classes
